GCP/requirements.txt
PyYAML==5.3.1
requests-oauthlib==1.3.0
tqdm==4.51.0
pandas==1.1.3
gspread==3.6.0
oauth2client==4.1.3
